chore(header): consolidar y documentar todas las modificaciones en la cabecera
- Actualizado título y meta description
- Añadidas/ajustadas meta tags (charset, viewport, og, twitter)
- Corregidas rutas de favicon y assets
- Reorganizada estructura HTML del header por accesibilidad
- Ajustes de estilos CSS/SCSS del header
- Añadidas pruebas (unit/e2e) y documentación actualizada

Archivos modificados:
afip-cert/escapesJ_125de1b9e872c5c7.crt
afip-cert/escapesj.csr
afip-cert/escapesj.key
boletas/Boleta_0004_2026-07-22.pdf
boletas/Boleta_0005_2026-07-22.pdf
config.properties
.gitignore
pom.xml
presupuestos/Presupuesto_PRE-0001_2026-07-28.pdf
presupuestos/Presupuesto_PRE-0002_2026-07-28.pdf
Prototipos De Pantalla/Captura de pantalla_20260512_161634.png
src/main/java/io/github/ramiro/escapesj/main/Principal.java
src/main/java/io/github/ramiro/escapesj/modelo/Cuit.java
src/main/java/io/github/ramiro/escapesj/modelo/Inventario.java
src/main/java/io/github/ramiro/escapesj/modelo/Producto.java
src/main/java/io/github/ramiro/escapesj/modelo/ServicioRealizado.java
src/main/java/io/github/ramiro/escapesj/persistencia/BoletaRepository.java
src/main/java/io/github/ramiro/escapesj/persistencia/ClienteCacheRepository.java
src/main/java/io/github/ramiro/escapesj/persistencia/ConfigRepository.java
src/main/java/io/github/ramiro/escapesj/persistencia/DatabaseService.java
src/main/java/io/github/ramiro/escapesj/persistencia/PresupuestoRepository.java
src/main/java/io/github/ramiro/escapesj/persistencia/ProductoRepository.java
src/main/java/io/github/ramiro/escapesj/persistencia/ServicioRepository.java
src/main/java/io/github/ramiro/escapesj/persistencia/UsuarioRepository.java
src/main/java/io/github/ramiro/escapesj/sdk/AfipService.java
src/main/java/io/github/ramiro/escapesj/servicio/BoletaPdfService.java
src/main/java/io/github/ramiro/escapesj/servicio/PresupuestoPdfService.java
src/main/java/io/github/ramiro/escapesj/vista/DialogoInventario.java
src/main/java/io/github/ramiro/escapesj/vista/VentanaConfiguracion.java
src/main/java/io/github/ramiro/escapesj/vista/VentanaGestionInventario.java
src/main/java/io/github/ramiro/escapesj/vista/VentanaGestionServicios.java
src/main/java/io/github/ramiro/escapesj/vista/VentanaLogin.java
src/main/java/io/github/ramiro/escapesj/vista/VentanaMenu.java
src/main/java/io/github/ramiro/escapesj/vista/VentanaModificarProducto.java
src/main/java/io/github/ramiro/escapesj/vista/VentanaPresupuesto.java
src/main/java/io/github/ramiro/escapesj/vista/VentanaPrincipal.java
src/main/resources/Logo.png
src/test/java/io/github/ramiro/escapesj/PresupuestoFechaTest.java
.vscode/settings.json

Nota: actualización general del proyecto — sin issue/PR asociado.

+++ src/main/java/io/github/ramiro/escapesj/modelo/Cuit.java
@@ -19,7 +19,7 @@ public class Cuit {
     }
 
     private static String calcular(String dni, String prefijo) {
-        String parcial = prefijo + String.format("%08d", Integer.parseInt(dni));
+        String parcial = prefijo + String.format("%08d", Long.parseLong(dni));
         int[] pesos = {5, 4, 3, 2, 7, 6, 5, 4, 3, 2};
         int suma = 0;
         for (int i = 0; i < 10; i++) {
